Reverted changes to fix crash on scroll down.  Changed return type of jbxvt_get_scroll_size() to int16_t.

+++ scroll.c
@@ -94,17 +94,15 @@ static void add_scroll_history(void)
 {
 	struct JBXVTScreen * s = jbxvt_get_current_screen();
 	enum { SIZE = sizeof(struct JBXVTLine) };
-	saved_lines = realloc(saved_lines,
-		(size_t)(++scroll_size * SIZE));
+	saved_lines = realloc(saved_lines, ++scroll_size * SIZE);
 	jb_require(saved_lines, "Cannot allocate memory"
 		" for scroll history");
 	// - 1 for index instead of size
 	memcpy(&saved_lines[scroll_size - 1], &s->line[s->cursor.y], SIZE);
 	trim();
 }
-// Returns the index into the source line array.
-static uint8_t copy_lines(const uint8_t i, const uint8_t j, const int8_t mod,
-	const uint8_t count)
+static int8_t copy_lines(const int8_t i, const int8_t j, const int8_t mod,
+	const int8_t count)
 {
 	if (i >= count)
 		return j;
@@ -145,7 +143,7 @@ static void sc_dn(xcb_connection_t * xc, const uint8_t row1, const
 	uint8_t row2, const int16_t count)
 {
 	struct JBXVTScreen * s = jbxvt_get_current_screen();
-	for(uint8_t j = copy_lines(0, row2, -1, count); j >= row1; --j)
+	for(int8_t j = copy_lines(0, row2, -1, count); j >= row1; --j)
 		move_line(j, count, s->line);
 	sc_common(xc, row1, row2, count, false);
 }
@@ -155,7 +153,7 @@ static void sc_up(xcb_connection_t * xc, const uint8_t row1, const
 	struct JBXVTScreen * s = jbxvt_get_current_screen();
 	if (s == jbxvt_get_screen_at(0) && row1 == 0)
 		add_scroll_history();
-	for(uint8_t j = copy_lines(0, row1, 1, count); j < row2; ++j)
+	for(int8_t j = copy_lines(0, row1, 1, count); j < row2; ++j)
 		move_line(j, -count, s->line);
 	sc_common(xc, row1, row2, count, true);
 }
